Set up file routes with lazy-loaded route modules
Introduce src/routes/ directory with route module files that follow
React Router's convention (Component, loader, action, ErrorBoundary
exports). Update routes.tsx to use lazy() imports for automatic code
splitting and consolidate RequireAuth into a layout route.
