+++ web/src/routes/feed.tsx
@@ -1,0 +1,1 @@
+export { Feed as Component, feedLoader as loader } from "@/pages/Feed";
