+++ web/src/routes/login.tsx
@@ -1,0 +1,1 @@
+export { Login as Component } from "@/pages/Login";
