+++ web/src/routes/upload.tsx
@@ -1,0 +1,1 @@
+export { Upload as Component, uploadAction as action } from "@/pages/Upload";
